Rename Manifest.json to manifest.json🔧
Greatest Indie dev to ever live... 🤦‍♂️ 

Fixed capitalization of Manifest.json, which is supposed to be manifest.json, not Manifest.json which is what it was, so I had to change the name to manifest.json so Thunderstore would be able to find a manifest.json in the archive. Yeah I made sure to draw that out to see how many times I could say Manifest.json.